### 2019-05-04 #1428 @whedon
[PRE REVIEW]: sluRm: A lightweight wrapper for HPC with Slurm
**Submitting author:** @gvegayon (<a href="http://orcid.org/0000-0002-3171-0844">George Vega Yon</a>)
**Repository:** <a href="https://github.com/USCbiostats/sluRm" target ="_blank">https://github.com/USCbiostats/sluRm</a>
**Version:** 0.1-0
