### 2019-05-28 #1428 @HenrikBengtsson
Sorry, I won't have the time to review this one any time soon, especially since I don't have access to a Slurm environment.  I think whoever reviews this should be able to test it and play with it using Slurm.  On a related note (request to the community), a Linux container with R and Slurm pre-configured would be useful (I would be interested in that outside of this submission).

My suggestion for alternative reviewers are:

 * Michel Lang (@mllg), author of [batchtools](https://cran.r-project.org/package=batchtools)
 * Michael Schubert (@mschubert), author of [clustermq](https://cran.r-project.org/package=clustermq)
 * Will Landau (@wlandau-lilly), author of [drake](https://github.com/ropensci/drake) and who I think have mentioned he got access to a Slurm environment

PS. Does "[PRE REVIEW]" in the subject mean "looking for reviewers" whereas "[REVIEW]" means "under review"?
