### 2019-05-28 #1428 @mschubert
I'd be happy to review this.<sup>1</sup>

However, could the editors please confirm that the advantages listed in the article (fewer dependencies, slightly different API) are sufficient as per JOSS guidelines?

> The software should be a significant contribution to the available open source software that either enables some new research challenges to be addressed or makes addressing research challenges significantly better (e.g., faster, easier, simpler)

<sup>1</sup> COI: author of package with similar functionality
