### 2019-05-29 #1428 @wlandau
I am willing to stay in the loop as a backup, but I am about to travel to a conference, so I would not be able to begin until next week. Also, the Slurm installation Henrik mentioned was experimental. It was all emulated on my local machine (including the one and only compute node) and I no longer currently have it.
