### 2019-05-29 #1428 @mllg
I can also review this and test the package on multiple slurm systems.
