### 2019-06-03 #1428 @karthik
Thanks Henrik for the reviewer suggestions and the helpful comments.

> However, could the editors please confirm that the advantages listed in the article (fewer dependencies, slightly different API) are sufficient as per JOSS guidelines?

@mschubert Good point.  @gvegayon: Can you please tell us a bit more about how your package is a significant improvement over existing tools?
