### 2019-06-04 4dce67d
sapply, read/write slurm_job and hooks for collection
