+++ tests/testthat/test-bash.R
@@ -5,7 +5,7 @@ test_that("If no slurm, then errors", {
   skip_if(slurm_available())
 
   tmp <- tempdir()
-  x <- suppressWarnings(Slurm_EvalQ(1+1, submit = FALSE, job_path = tmp))
+  x <- suppressWarnings(Slurm_EvalQ(1+1, plan = "none", job_path = tmp))
 
   expect_error(scancel(x), "not available")
   expect_error(squeue(x), "not available")
