### 2019-06-05 cc00a5a
slurm.conf and adding some examples for the batch wrappers
