+++ R/bash.R
@@ -1,10 +1,13 @@
-#' Check whether Slurm is available in the system
-#' @return A Logical scalar indicating whether Slurm is available (`TRUE`) or not
-#' (`FALSE`)
+#' @details The function `slurm_available` checks whether Slurm is available in
+#' the system or not. It is usually called before calling any bash wrapper.
+#' If available, the function will return `TRUE`, otherwise `FALSE`.
+#'
 #' @examples
+#' # Are we under a Slurm Cluster?
 #' slurm_available()
 #'
 #' @export
+#' @rdname sbatch
 slurm_available <- function() {
 
   x <- tryCatch(
@@ -24,6 +27,17 @@ stopifnot_slurm <- function() {
 
   if (!slurm_available())
     stop("Slurm is not available on this system.", call. = FALSE)
+
+}
+
+stopifnot_submitted <- function(x) {
+
+  if (is.na(x)) {
+    stop(
+      "This job hasn't started/been submitted yet. Nothing to do.",
+      call. = FALSE
+    )
+  }
 
 }
 
@@ -50,15 +64,61 @@ silent_system2 <- function(...) {
 
 }
 
-which_submit <- function(job, what) {
+#' This environment sets and gets the latest submitted job. The function
+#' [last_submitted_job] is a wrapper of it visible to the user.
+#' @noRd
+LAST_SUBMITTED_JOB <- (function() {
 
-  state(job)
+  record <- new.env(parent = emptyenv())
+
+  record$job <- NULL
+  record$set <- function(job) {
+
+    if (!inherits(job, "slurm_job"))
+      stop("The `job` argument must be an object of class `slurm_job`.",
+           call. = FALSE)
+
+    record$job <- job
+
+    invisible()
+  }
+  record$get <- function() {
+    record$job
+  }
+
+  return(record)
+
+})()
+
+#' @rdname sbatch
+#' @export
+#' @details The `las_submitted_job` function will return the latest `slurm_job`
+#' object that was submitted via [sbatch] in the current session. The `last_job`
+#' function is just an alias of the later. If no job has been submitted, then
+#' the resulting value will be `NULL`.
+#' @examples
+#' \dontrun{
+#' # The last_job function can be handy when `plan = "collect"` in a called,
+#' # for example
+#' job <- Slurm_lapply(1:1000, function(i) runif(100), njobs = 2, plan = "collect")
+#'
+#' # Post collection analysis
+#' state(last_job())
+#' }
+last_submitted_job <- function() {
+
+  LAST_SUBMITTED_JOB$get()
 
 }
 
+#' @export
+#' @rdname slurm_job
+last_job <- last_submitted_job
+
 #' R wrappers for ommands included in *Slurm*
 #'
-#' @param x An object of class `slurm_job`.
+#' @param x Either an object of class `slurm_job`, or, in some cases, an
+#' integer as a Slurm jobid. Note that some functions allow passing no arguments.
 #' @param wait Logical scalar. When `TRUE` the function will pass the `--wait`
 #' flag to `Slurm` and set `wait = TRUE` in the [system2] call.
 #' @param submit Logical, when `TRUE` calls [sbatch] to submit the job to slurm.
@@ -71,7 +131,7 @@ which_submit <- function(job, what) {
 #'
 #' @export
 #' @aliases submit
-sbatch <- function(x, wait=TRUE, submit = TRUE, ...) UseMethod("sbatch")
[…]
