+++ tests/testthat/test-bash.R
@@ -7,10 +7,15 @@ test_that("If no slurm, then errors", {
   tmp <- tempdir()
   x <- suppressWarnings(Slurm_EvalQ(1+1, plan = "none", job_path = tmp))
 
-  expect_error(scancel(x), "not available")
-  expect_error(squeue(x), "not available")
+  expect_error(scancel(1), "not available")
+  expect_error(squeue(1), "not available")
   expect_error(sbatch(x), "not available")
-  expect_error(sacct(x), "not available")
+  expect_error(sacct(1), "not available")
+
+  expect_error(scancel(x), "started")
+  expect_error(squeue(x), "started")
+  expect_error(sacct(x), "started")
+
 
 })
 
