+++ man/slurm_job.Rd
@@ -1,6 +1,7 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/slurm_job-class.R
-\name{slurm_job}
+% Please edit documentation in R/bash.R, R/slurm_job-class.R
+\name{last_job}
+\alias{last_job}
 \alias{slurm_job}
 \alias{new_slurm_job}
 \alias{print.slurm_job}
@@ -8,6 +9,8 @@
 \alias{write_slurm_job}
 \title{Creating Slurm jobs}
 \usage{
+last_job()
+
 new_slurm_job(call, rscript, bashfile, robjects, njobs, job_opts,
   hooks = NULL)
 
