+++ man/sbatch.Rd
@@ -1,38 +1,54 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/bash.R
-\name{sbatch}
+\name{slurm_available}
+\alias{slurm_available}
+\alias{last_submitted_job}
 \alias{sbatch}
 \alias{submit}
 \alias{sbatch.slurm_job}
 \alias{squeue}
+\alias{squeue.default}
 \alias{squeue.slurm_job}
 \alias{scancel}
+\alias{scancel.default}
 \alias{scancel.slurm_job}
 \alias{sacct}
-\alias{sacct.slurm_job}
 \alias{sacct.default}
+\alias{sacct.slurm_job}
+\alias{slurm.conf}
 \title{R wrappers for ommands included in \emph{Slurm}}
 \usage{
+slurm_available()
+
+last_submitted_job()
+
 sbatch(x, wait = TRUE, submit = TRUE, ...)
 
 \method{sbatch}{slurm_job}(x, wait = TRUE, submit = TRUE, ...)
 
-squeue(x, ...)
+squeue(x = NULL, ...)
+
+\method{squeue}{default}(x = NULL, ...)
 
 \method{squeue}{slurm_job}(x, ...)
 
 scancel(x, ...)
 
+\method{scancel}{default}(x, ...)
+
 \method{scancel}{slurm_job}(x, ...)
 
 sacct(x, ...)
 
+\method{sacct}{default}(x, brief = TRUE, parsable = TRUE, ...)
+
 \method{sacct}{slurm_job}(x, ...)
 
-\method{sacct}{default}(x, brief = TRUE, parsable = TRUE, ...)
+slurm.conf()
 }
 \arguments{
-\item{x}{An object of class \code{slurm_job}.}
+\item{x}{Either an object of class \code{slurm_job}, or, in some cases, an
+integer as a Slurm jobid. Note that some functions allow passing no arguments.}
 
 \item{wait}{Logical scalar. When \code{TRUE} the function will pass the \code{--wait}
 flag to \code{Slurm} and set \code{wait = TRUE} in the \link{system2} call.}
@@ -48,7 +64,42 @@ to the command line function \code{sacct}.}
 R wrappers for ommands included in \emph{Slurm}
 }
 \details{
+The function \code{slurm_available} checks whether Slurm is available in
+the system or not. It is usually called before calling any bash wrapper.
+If available, the function will return \code{TRUE}, otherwise \code{FALSE}.
+
+The \code{las_submitted_job} function will return the latest \code{slurm_job}
+object that was submitted via \link{sbatch} in the current session. The \code{last_job}
+function is just an alias of the later. If no job has been submitted, then
+the resulting value will be \code{NULL}.
+
 In the case of \code{sbatch}, function takes an object of class \code{slurm_job} and
 submits it to the queue. In debug mode the job will be submitted via \code{sh}
 instead.
+
+The function \code{slurm.conf} is a wrapper of the function \code{scontrol} that
+returns configuration info about Slurm, in particular, the underlaying command
+that is called is \code{scontrol show conf}. This returns a named character vector
+with configuration info about the cluster. The name of this function matches
+the name of the file that holds this information.
+}
+\examples{
+# Are we under a Slurm Cluster?
+slurm_available()
+
+\dontrun{
+# The last_job function can be handy when `plan = "collect"` in a called,
+# for example
+job <- Slurm_lapply(1:1000, function(i) runif(100), njobs = 2, plan = "collect")
+
+# Post collection analysis
+state(last_job())
+}
+
+\dontrun{
+# allows?
+sconfig <- slurm.conf() # We first retrieve the info.
+sconfig["MaxArraySize"]
+}
+
 }
