+++ README.md
@@ -42,7 +42,7 @@ And the development version from [GitHub](https://github.com/) with:
 devtools::install_github("USCbiostats/sluRm")
 ```
 
-## Examples
+## Example 1: Computing means (and looking under the hood)
 
 ``` r
 library(sluRm)
@@ -59,22 +59,23 @@ x <- replicate(100, runif(50), simplify = FALSE)
 We can use the function `Slurm_lapply` to distribute computations
 
 ``` r
-ans <- Slurm_lapply(x, mean, submit = FALSE)
+ans <- Slurm_lapply(x, mean, plan = "none")
 #  Warning: [submit = FALSE] The job hasn't been submitted yet. Use sbatch() to submit the job, or you can submit it via command line using the following:
-#  sbatch --chdir=/home/vegayon/Documents/sluRm --job-name=sluRm-job-74e7b8110a3 /home/vegayon/Documents/sluRm/sluRm-job-74e7b8110a3/01-bash.sh
+#  sbatch --chdir=/home/vegayon/Documents/sluRm --job-name=sluRm-job-71835505ddb0 /home/vegayon/Documents/sluRm/sluRm-job-71835505ddb0/01-bash.sh
 Slurm_clean(ans) # Cleaning after you
 ```
 
-Notice the `submit = FALSE` option. To get more info, we can actually
-set the verbose mode on
+Notice the `plan = "none"` option, this tells `Slurm_lapply` to only
+create the job object, but do nothing with it, i.e., skip submission. To
+get more info, we can actually set the verbose mode on
 
 ``` r
 opts_sluRm$verbose_on()
-ans <- Slurm_lapply(x, mean, submit = FALSE)
+ans <- Slurm_lapply(x, mean, plan = "none")
 #  
 #  --------------------------------------------------------------------------------
 #  [VERBOSE MODE ON] The R script that will be used is located at:
-#  /home/vegayon/Documents/sluRm/sluRm-job-74e7b8110a3/00-rscript.r
+#  /home/vegayon/Documents/sluRm/sluRm-job-71835505ddb0/00-rscript.r
 #  and has the following contents:
 #  --------------------------------------------------------------------------------
 #  .libPaths(c("/usr/local/lib/R/site-library", "/usr/lib/R/site-library", "/usr/lib/R/library"))
@@ -87,39 +88,60 @@ ans <- Slurm_lapply(x, mean, submit = FALSE)
 #      y
 #  }
 #  ARRAY_ID         <- as.integer(Slurm_env("SLURM_ARRAY_TASK_ID"))
-#  INDICES          <- readRDS("/home/vegayon/Documents/sluRm/sluRm-job-74e7b8110a3/INDICES.rds")
-#  X                <- readRDS(sprintf("/home/vegayon/Documents/sluRm/sluRm-job-74e7b8110a3/X_%04d.rds", ARRAY_ID))
-#  FUN              <- readRDS("/home/vegayon/Documents/sluRm/sluRm-job-74e7b8110a3/FUN.rds")
-#  mc.cores         <- readRDS("/home/vegayon/Documents/sluRm/sluRm-job-74e7b8110a3/mc.cores.rds")
-#  seeds            <- readRDS("/home/vegayon/Documents/sluRm/sluRm-job-74e7b8110a3/seeds.rds")
+#  INDICES          <- readRDS("/home/vegayon/Documents/sluRm/sluRm-job-71835505ddb0/INDICES.rds")
+#  X                <- readRDS(sprintf("/home/vegayon/Documents/sluRm/sluRm-job-71835505ddb0/X_%04d.rds", ARRAY_ID))
[…]
