### 2019-06-05 9f36f07
scancel can be called without arguments (as it should). Also, added nicer example of resubmitting failed jobs.
