+++ README.md
@@ -1,50 +1,42 @@
 
-[![Travis build
-status](https://travis-ci.org/USCbiostats/sluRm.svg?branch=master)](https://travis-ci.org/USCbiostats/sluRm)
-[![codecov](https://codecov.io/gh/USCbiostats/sluRm/branch/master/graph/badge.svg)](https://codecov.io/gh/USCbiostats/sluRm)
-[![lifecycle](https://img.shields.io/badge/lifecycle-experimental-orange.svg)](https://www.tidyverse.org/lifecycle/#experimental)
+[![Travis build status](https://travis-ci.org/USCbiostats/sluRm.svg?branch=master)](https://travis-ci.org/USCbiostats/sluRm) [![codecov](https://codecov.io/gh/USCbiostats/sluRm/branch/master/graph/badge.svg)](https://codecov.io/gh/USCbiostats/sluRm) [![lifecycle](https://img.shields.io/badge/lifecycle-experimental-orange.svg)](https://www.tidyverse.org/lifecycle/#experimental)
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
-# sluRm: A lightweight wrapper for Slurm <img src="man/figures/logo.png" height="180px" align="right"/>
 
-Slurm Workload Manager is a popular HPC cluster job scheduler found in
-many of the top 500 super computers. The `sluRm` R package provides an R
-wrapper to it that matches the parallel package’s syntax, this is, just
-like `parallel` provides the `parLapply`, `parMap`, `parSapply`, etc.,
-`sluRm` provides `Slurm_lapply`, `Slurm_Map`, `Slurm_sapply`, etc.
 
-While there are other alternatives such as `future.batchtools`,
-`batchtools`, `clustermq`, and `rslurm`, this R package has the
-following goals:
+
+sluRm: A lightweight wrapper for Slurm <img src="man/figures/logo.png" height="180px" align="right"/>
+=====================================================================================================
+
+Slurm Workload Manager is a popular HPC cluster job scheduler found in many of the top 500 super computers. The `sluRm` R package provides an R wrapper to it that matches the parallel package's syntax, this is, just like `parallel` provides the `parLapply`, `parMap`, `parSapply`, etc., `sluRm` provides `Slurm_lapply`, `Slurm_Map`, `Slurm_sapply`, etc.
+
+While there are other alternatives such as `future.batchtools`, `batchtools`, `clustermq`, and `rslurm`, this R package has the following goals:
 
 1.  It is dependency free, which means that it works out-of-the-box
 
-2.  Puts an emphasis on been similar to the workflow in the R package
-    `parallel`
+2.  Puts an emphasis on been similar to the workflow in the R package `parallel`
 
-3.  It provides a general framework for the user to create its own
-    wrappers without using template files.
+3.  It provides a general framework for the user to create its own wrappers without using template files.
 
-4.  Is specialized on Slurm, meaning more flexibility (no need to modify
-    template files), and, in the future, better debuging tools (e.g. job
-    resubmission).
[…]
