### 2019-06-05 #1428 @gvegayon
Thanks all (@mschubert @mllg @wlandau @HenrikBengtsson) for all the comments/suggestions! @karthik, besides of it be lightweight (actually, dependency free), the other important contributions are:

* Tailored for Slurm workflow (no need to use template files to set up a job, this is done on the fly).

* Accessing bash commands directly from within R (cancel jobs, query jobs, see details on the job account, submitting jobs, etc.)

```r
# Canceling a previously created job
scancel(job)

# Canceling all my jobs in the cluster
scancel(user = "vegayon")

# Canceling a given job id
scancel(9291231)
```

* smoothly resubmitting jobs, or parts of it (which is something some of my colleagues value a lot), e.g.:
```r
library(sluRm)

# A simple expr evaluation, WhoAmI() gives some info about the node
x <- Slurm_EvalQ(sluRm::WhoAmI(), njobs = 20, plan = "wait")

# Suppose jobs 1, 2, and 15-20 failed, then we can do the following
sbatch(x, array = "1,2,15-20")

# And, if status OK, then collect safely the entire array
ans <- Slurm_collect(x)
```

This last point is critical (I see partial failure all the time, and is very common in heterogeneous clusters). And I'm not sure how easy/complicated is to resubmit parts of jobs in the other frameworks, or how easy/complicated is to add these features to the other frameworks.

I tried to work out a comparison table with some of the tools mentioned [here](https://github.com/USCbiostats/sluRm/#vs) (tried to fill it to the best of my understanding, would happily receive comments/complains/suggestions from any of you :) ).

LMK what are your thoughts

cc my co-author @pmarjora
