### 2019-06-06 #1428 @karthik
Thank you @gvegayon for addressing this issue.  This meets JOSS' criteria and I'm happy to send this out for review. Thanks @mschubert and @mllg for volunteering. I'll assign you shortly.
