### 2019-06-06 #1493 @whedon
[REVIEW]: sluRm: A lightweight wrapper for HPC with Slurm
**Submitting author:** @gvegayon (<a href="http://orcid.org/0000-0002-3171-0844">George Vega Yon</a>)
**Repository:** <a href="https://github.com/USCbiostats/sluRm" target ="_blank">https://github.com/USCbiostats/sluRm</a>
**Version:** v0.2-0

- [x] **Version:** v0.2-0

- [x] **Version:** v0.2-0
