### 2019-06-06 87a0f97
Adding badge and DOI id for Slurm paper [skip ci]
