+++ paper/paper.bib
@@ -7,14 +7,22 @@
     url = {https://www.R-project.org/},
   }
 
-@INPROCEEDINGS{Jette02slurm:simple,
-    author = {Morris A. Jette and Andy B. Yoo and Mark Grondona},
-    title = {SLURM: Simple Linux Utility for Resource Management},
-    booktitle = {In Lecture Notes in Computer Science: Proceedings of Job Scheduling Strategies for Parallel Processing (JSSPP) 2003},
-    year = {2002},
-    pages = {44--60},
-    publisher = {Springer-Verlag}
+@inproceedings{Jette02slurm,
+abstract = {A new cluster resource management system called Simple Linux Utility Resource Management (SLURM) is described in this paper. SLURM, initially developed for large Linux clusters at the Lawrence Livermore National Laboratory (LLNL), is a simple cluster manager that can scale to thousands of processors. SLURM is designed to be flexible and fault-tolerant and can be ported to other clusters of different size and architecture with minimal effort. We are certain that SLURM will benefit both users and system architects by providing them with a simple, robust, and highly scalable parallel job execution environment for their cluster system.},
+address = {Berlin, Heidelberg},
+author = {Yoo, Andy B and Jette, Morris A and Grondona, Mark},
+booktitle = {Job Scheduling Strategies for Parallel Processing},
+doi = {10.1007/10968987_3},
+editor = {Feitelson, Dror and Rudolph, Larry and Schwiegelshohn, Uwe},
+isbn = {978-3-540-39727-4},
+pages = {44--60},
+publisher = {Springer Berlin Heidelberg},
+title = {{SLURM: Simple Linux Utility for Resource Management}},
+url = {http://link.springer.com/10.1007/10968987{\_}3},
+year = {2003}
 }
+
+
 
 @Manual{Marchand2018,
     year   = {2017},
