### 2019-06-08 USCbiostats/sluRm#5 @mschubert
[JOSS review] Travis failing on latest master
This is a part of the JOSS review outlined in https://github.com/openjournals/joss-reviews/issues/1493.

https://travis-ci.org/USCbiostats/sluRm/branches

Can you please fix your tests?
