### 2019-06-08 USCbiostats/sluRm#6 @mschubert
[JOSS review] Inconsistent authors
This is a part of the JOSS review outlined in https://github.com/openjournals/joss-reviews/issues/1493.

The authors on the paper and in DESCRIPTION are different. Is there any reason for that?

Looking at  the commit history, @pmarjora contributed mostly on improving documentation. This is fine, but can you please confirm your authorship and that you are willing to take responsibility for the paper and the software (insofar as this is usually the case for co-authors on publications)?
