### 2019-06-08 USCbiostats/sluRm#7 @mschubert
[JOSS review] Update paper to reflect pre-review comments
This is a part of the JOSS review outlined in https://github.com/openjournals/joss-reviews/issues/1493.

You gave an updated view of the package's strengths in pre-review:
https://github.com/openjournals/joss-reviews/issues/1428#issuecomment-499282135.

It would be nice it this was better reflected in the paper (i.e., just updated to include your points - e.g. resubmit on failure)
