### 2019-06-08 USCbiostats/sluRm#8 @mschubert
[JOSS review] Citations in paper
This is a part of the JOSS review outlined in https://github.com/openjournals/joss-reviews/issues/1493.

The citation for `batchtools` is wrong: this is for `BatchJobs`, rather use this:

http://joss.theoj.org/papers/10.21105/joss.00135
