### 2019-06-08 USCbiostats/sluRm#8 @mschubert
In addition, it would be nice if you could cite `clustermq` as one of the related tools (note my COI here, but I think this is still relevant)
