### 2019-06-10 USCbiostats/sluRm#8 @mschubert
Also, please include DOIs in the references to satisfy the checklist

> References: Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?
