### 2019-06-10 USCbiostats/sluRm#9 @mschubert
[JOSS review] [bug?] Tempfiles overwritten with 2 job objects
This is a part of the JOSS review outlined in https://github.com/openjournals/joss-reviews/issues/1493.

Consider the following:

```r
job1 <- Slurm_EvalQ(sluRm::WhoAmI(), njobs = 1, plan = "submit")
job2 <- Slurm_EvalQ(sluRm::WhoAmI(), njobs = 1, plan = "submit")

Slurm_clean(job1)
Slurm_collect(job2)
# Error: Nothing to retrieve. (see ?status).
```

I assume this is not intended behaviour.
