### 2019-06-10 USCbiostats/sluRm#10 @mschubert
[JOSS review] Chunking of jobs
This is a part of the JOSS review outlined in https://github.com/openjournals/joss-reviews/issues/1493.

As far as I can tell, individual function calls are always submitted as array jobs.

What if you have millions of calls?
Does that mean `sluRm` will submit an array with millions of entries?

In other words, can `sluRm` chunk together multiple function calls in one jobs?

I assume it can not. This is fine, but should be documented.
