### 2019-06-10 USCbiostats/sluRm#11 @mschubert
[JOSS review] Jobs will fail if $(pwd) is not on network storage
This is a part of the JOSS review outlined in https://github.com/openjournals/joss-reviews/issues/1493.

Temporary files are created in the current working directory.

If these are not accessible from the network, jobs will fail.

This should be documented.
