### 2019-06-10 USCbiostats/sluRm#10 @mschubert
Looking at e.g. `Slurm_lapply`, it seems I was wrong and chunking is supported :+1:

Please document!
