### 2019-06-10 USCbiostats/sluRm#12 @mschubert
[JOSS review] [bug?] Submitting only one `lapply` job fails
This is a part of the JOSS review outlined in https://github.com/openjournals/joss-reviews/issues/1493.

The following fails:

```r
library(sluRm)
Slurm_lapply(1, function(x) x*2)
# Error in splits[[j]] : subscript out of bounds
# In addition: Warning message:
# `X` is not a list. The function will coerce it into one using `as.list`
```

Related: Internal parallelism via `mc.cores` defaults to `2`, which wastes a core if there are not more calls than jobs.
