### 2019-06-10 USCbiostats/sluRm#13 @mschubert
[JOSS review] Community guidelines
This is a part of the JOSS review outlined in https://github.com/openjournals/joss-reviews/issues/1493.

The JOSS check list states:

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

However, the only contributing guidelines are a linked CoC.

Can you add a sentence on how to report issues/PRs/seek support? (e.g. should users open an issue, email the author, or send a carrier pigeon)
