### 2019-06-10 USCbiostats/sluRm#7 @mschubert
Related: also include shortcomings vs. other packages.

> Compared to rslurm (Marchand, Smorul, & Carrol, 2017), this R package’s work-
flow is closer to the parallel package

I don't see how this is a main advantage your package. You can either have one apply function or a family, that doesn't make one better than the other.

Also referenced parallel functions are wrong:

```r
> parallel::parMap
Error: 'parMap' is not an exported object from 'namespace:parallel'
```

> Compared to batchtools

Maybe in-file chunking is more important than the automatically installed dependencies. I also like the zero-config ready-to-go approach. Have you thought about mentioning those?

> and to future.batchtools

Downside of `sluRm` is also that it doesn't support e.g. parallel `foreach` for providing paralellism to other packages. Make differences clear, don't just pretend yours is better.

> In terms of flexibility, sluRm’s API is significantly more flexible compared to the
other R packages mentioned.

Well, yes, for using slurm only. It's less flexible in terms of schedulers. Also, "significantly" is hard to quantify so I'd prefer you drop that word.
