### 2019-06-10 #1493 @mschubert
I finished my first 2 passes. Nice work @gvegayon and @pmarjora!

I'm happy for this to be accepted if the following issues are addressed: [#6](https://github.com/USCbiostats/sluRm/issues/6), [#7](https://github.com/USCbiostats/sluRm/issues/7), [#8](https://github.com/USCbiostats/sluRm/issues/8), [#13](https://github.com/USCbiostats/sluRm/issues/13) (others are optional, but would be nice to have)
