### 2019-06-10 #1493 @gvegayon
@mschubert this is great, we really appreciate it! We will address those (and the optional issues) ASAP. Thanks!
