### 2019-06-11 0bd9f03
Addressing issues #10 #11 and #12
