+++ tests/testthat/test-map.R
@@ -31,7 +31,7 @@ test_that("Exporting", {
   x <- tempdir()
   suppressMessages({
     opts_sluRm$set_chdir(x)
-    opts_sluRm$set_job_name("map-2")
+    opts_sluRm$set_job_name("map-3")
     opts_sluRm$debug_on()
     opts_sluRm$verbose_off()
   })
@@ -48,5 +48,21 @@ test_that("Exporting", {
   sol <- Slurm_collect(ans)
 
   expect_equal(sol[[1]], list(x=a[[1]], y=b[[1]]))
+
+})
+
+
+test_that("Warnings and errors", {
+
+  x <- tempdir()
+  suppressMessages({
+    opts_sluRm$set_chdir(x)
+    opts_sluRm$set_job_name("map-warns-and-errors")
+    opts_sluRm$debug_on()
+    opts_sluRm$verbose_off()
+  })
+
+  expect_warning(Slurm_Map(function(x,y) mean(x,y), x = 1, y = 1:2, njobs = 4, plan = "none"), "length")
+  expect_error(Slurm_Map(function(x) mean(x), x = list(1), 4, njobs = 2, plan = "none"), "unname")
 
 })
