### 2019-06-11 USCbiostats/sluRm#12 @gvegayon
Thanks. This should be fixed now. Regarding the default of 2 mc.cores, that's not an issue since `mclapply` automatically checks for this. If length(X) == 1, then mc.cores defaults to `lapply` instead. I was just trying to be consistent with what is in `parallel`.
