### 2019-06-11 USCbiostats/sluRm#12 @gvegayon
Also, this was a nasty bug. Thanks for flagging that!
