### 2019-06-11 USCbiostats/sluRm#11 @gvegayon
Added more documentation on this.
