### 2019-06-11 USCbiostats/sluRm#10 @gvegayon
Yes! Of course, if that were not the case, then ppl should go to https://github.com/mschubert/clustermq instead :). Documented.
