### 2019-06-11 USCbiostats/sluRm#12 @mschubert
> Regarding the default of 2 mc.cores, that's not an issue since `mclapply` automatically checks for this. If length(X) == 1, then mc.cores defaults to `lapply` instead.

That's true for mclapply in R, but not for the job that requests 2 cores (and only uses 1)
