### 2019-06-11 USCbiostats/sluRm#12 @gvegayon
OK, good point. Nevertheless, I'm honestly not sure regarding this as a
problem, since in most cases using Slurm_lapply/Slurm_Map/Slurm_sapply
users will be submitting a job with length(X) >> 1. I do need to get more
details about what happens behind scenes when you request more than you are
using, and also regarding how much enforced is the request of resources (in
the manual of sbatch, the "Advice the controller" phrase is a bit confusing
for me).

Either way, I follow your advice and will set the default to equal 1L.

[ ] Set `mc.cores = 1L` in `Slurm_lapply`/`Map`

George G. Vega Yon
+1 (626) 381 8171
https://ggvy.cl

On Tue, Jun 11, 2019 at 10:06 AM Michael Schubert <notifications@github.com>
wrote:

> Regarding the default of 2 mc.cores, that's not an issue since mclapply
> automatically checks for this. If length(X) == 1, then mc.cores defaults to
> lapply instead.
>
> That's true for mclapply in R, but not for the job that requests 2 cores
> (and only uses 1)
>
