### 2019-06-17 USCbiostats/sluRm#14 @mllg
[JOSS review] cli argument --chdir not portable
This is a part of the JOSS review outlined in openjournals/joss-reviews#1493.

`--chdir` is not available for all slurm installations.

```
$ sbatch --version
slurm 16.05.9

$ sbatch --chdir=~
sbatch: unrecognized option '--chdir=~'
Try "sbatch --help" for more information
```
`-D` seems to be a workaround.
