### 2019-06-17 USCbiostats/sluRm#15 @mllg
[JOSS review] Seeding
This is a part of the JOSS review outlined in openjournals/joss-reviews#1493.

The seeds currently default to `1:njobs`. I've followed a similar approach in `batchtools` where I determine a start seed randomly, and then increment the seed for each job. Even with a random initialization, I'm not sure if this approach is feasible (see https://github.com/mllg/batchtools/issues/81). Defaulting to `1:njobs` seems to be even worse, there is a reason RNGs are not initialized with a constant value.
