### 2019-06-17 USCbiostats/sluRm#16 @mllg
[JOSS review] Comparison table
This is a part of the JOSS review outlined in openjournals/joss-reviews#1493.

The comparison table in the readme says that `batchtools` cannot re-run jobs, but batchtools is perfectly capable of it:

```r
library(batchtools)
reg = makeRegistry(NA)
f = function(x) if (x == 3) stop(3) else x^2
btlapply(1:5, f, reg = reg)
# -> no result

# partial results are available in registry
reduceResultsList()

# re-submit failed jobs
# (job #3 will fail again of course, just for demonstration)
submitJobs(findErrors())
```
