### 2019-06-17 USCbiostats/sluRm#17 @mllg
[JOSS review] Prefer file.path()?
This is a part of the JOSS review outlined in openjournals/joss-reviews#1493.

This is nitpicking, but I'd suggest to use `file.path()` instead of `paste()` or `paste0()`. According to the documentation, `file.path()` should be faster.
Feel free to just close this issue if you disagree or have other reasons to not use `file.path()`.
