### 2019-06-17 USCbiostats/sluRm#18 @mllg
[JOSS review] Add a more realistic example
This is a part of the JOSS review outlined in openjournals/joss-reviews#1493.

It would be very important for the user to have a vignette with a more realistic example, especially on how to deal with failed jobs (retrieve logs, run job interactively to generate a traceback (is this possible?), re-submit subset of failed jobs).
