### 2019-06-17 #1493 @mllg
I've also finished my first pass. Issues are referenced here, #14 and #15 are mandatory.
