### 2019-06-19 8f15e50
Addressing #16
