+++ R/Map.R
@@ -39,7 +39,7 @@ Slurm_Map <- function(
   njobs       = 2L,
   mc.cores    = getOption("mc.cores", 2L),
   job_name    = opts_sluRm$get_job_name(),
-  job_path    = opts_sluRm$get_chdir(),
+  tmp_path    = opts_sluRm$get_tmp_path(),
   plan        = "collect",
   sbatch_opt  = list(ntasks=1L, `cpus-per-task`=mc.cores),
   rscript_opt = list(vanilla=TRUE),
@@ -91,7 +91,7 @@ Slurm_Map <- function(
   }
 
   # Setting the job name
-  opts_sluRm$set_chdir(job_path)
+  opts_sluRm$set_tmp_path(tmp_path)
   opts_sluRm$set_job_name(job_name)
 
   # Writing the data on the disk -----------------------------------------------
@@ -151,7 +151,8 @@ Slurm_Map <- function(
     bashfile = snames("sh"),
     robjects = NULL,
     njobs    = njobs,
-    job_opts = opts_sluRm$get_opts(),
+    opts_job = opts_sluRm$get_opts_job(),
+    opts_r   = opts_sluRm$get_opts_r(),
     hooks    = hooks
   )
 
