+++ R/options.R
@@ -1,11 +1,11 @@
 
 #' Get and set default options for `sbatch` and `sluRm` internals
 #'
-#' Most of the functions in the `sluRm` package use `chdir` and `job-name`
+#' Most of the functions in the `sluRm` package use `tmp_path` and `job-name`
 #' options to write and submit jobs to **Slurm**. These options have global
 #' defaults that are set and retrieved using `opts_sluRm`.
 #'
-#' Whatever the path specified on `chdir`, all nodes should have access to it.
+#' Whatever the path specified on `tmp_path`, all nodes should have access to it.
 #' Moreover, it is recommended to use a path located in a high-performing drive.
 #' See for example \url{https://hpcc.usc.edu/support/infrastructure/temporary-disk-space/}.
 #'
@@ -29,9 +29,9 @@
 #'
 #' Slurm options
 #'
-#' - `set_chdir : function (path, recursive = TRUE)` \Sexpr{attr(sluRm::opts_sluRm$set_chdir, "desc")}
+#' - `set_tmp_path : function (path, recursive = TRUE)` \Sexpr{attr(sluRm::opts_sluRm$set_tmp_path, "desc")}
 #'
-#' - `get_chdir : function ()` \Sexpr{attr(sluRm::opts_sluRm$get_chdir, "desc")}
+#' - `get_tmp_path : function ()` \Sexpr{attr(sluRm::opts_sluRm$get_tmp_path, "desc")}
 #'
 #' - `set_job_name : function (path, check = TRUE, overwrite = TRUE)` \Sexpr{attr(sluRm::opts_sluRm$set_job_name, "desc")}.
 #'
@@ -46,14 +46,16 @@
 #'
 #' - `set_opts : function (...)` \Sexpr{attr(sluRm::opts_sluRm$set_opts, "desc")}
 #'
-#' - `get_opts : function (...)` \Sexpr{attr(sluRm::opts_sluRm$get_opts, "desc")}
+#' - `get_opts_job : function (...)` \Sexpr{attr(sluRm::opts_sluRm$get_opts_job, "desc")}
+#'
+#' - `get_opts_r : function (...)` \Sexpr{attr(sluRm::opts_sluRm$get_opts_r, "desc")}
 #'
 #'
 #' @examples
 #'
 #' # Common setup
 #' \dontrun{
-#' opts_sluRm$set_chdir("/staging/pdt/vegayon")
+#' opts_sluRm$set_tmp_path("/staging/pdt/vegayon")
 #' opts_sluRm$set_job_name("simulations-1")
 #' opts_slurm$set_opts(partition="thomas", account="lc_pdt")
 #' }
@@ -63,47 +65,48 @@ opts_sluRm <- (function() {
 
   # Default chdir is null and will be set at the first call of the function
   OPTS_SLURM            <- new.env(parent = emptyenv())
-  OPTS_SLURM$chdir      <- NULL
+  # OPTS_SLURM$chdir      <- NULL
   OPTS_SLURM$`job-name` <- NULL
 
-  OPTS_R         <- new.env(parent = emptyenv())
-  OPTS_R$debug   <- FALSE
-  OPTS_R$cmd     <- "sbatch"
-  OPTS_R$verbose <- FALSE
+  OPTS_R          <- new.env(parent = emptyenv())
+  OPTS_R$tmp_path <- NULL
+  OPTS_R$debug    <- FALSE
+  OPTS_R$cmd      <- "sbatch"
+  OPTS_R$verbose  <- FALSE
 
 
   # JOB PATH -------------------------------------------------------------------
   # Function to set job path
-  set_chdir <- function(path, recursive = TRUE, overwrite = FALSE) {
[…]
