+++ tests/testthat/test-lapply.R
@@ -5,7 +5,7 @@ test_that("Works in debug mode", {
   # Setting it up
   x <- tempdir()
   suppressMessages({
-    opts_sluRm$set_chdir(x)
+    opts_sluRm$set_tmp_path(x)
     opts_sluRm$set_job_name("lapply-1")
     opts_sluRm$debug_on()
     opts_sluRm$verbose_off()
@@ -27,7 +27,7 @@ test_that("Exporting", {
   # Setting it up
   x <- tempdir()
   suppressMessages({
-    opts_sluRm$set_chdir(x)
+    opts_sluRm$set_tmp_path(x)
     opts_sluRm$set_job_name("lapply-2")
     opts_sluRm$debug_on()
     opts_sluRm$verbose_off()
@@ -51,7 +51,7 @@ test_that("Warnings and errors", {
 
   x <- tempdir()
   suppressMessages({
-    opts_sluRm$set_chdir(x)
+    opts_sluRm$set_tmp_path(x)
     opts_sluRm$set_job_name("lapply-warns-and-errors")
     opts_sluRm$debug_on()
     opts_sluRm$verbose_off()

[…]
