### 2019-06-24 USCbiostats/sluRm#17 @gvegayon
Overall, to build filepaths I use `sprintf` which is about the same as `file.path`. I do this in the function `snames` which is a general wrapper. I do see that I was still using `paste` in some places. Just changed that (thanks!)
