### 2019-06-24 USCbiostats/sluRm#16 @gvegayon
Didn't know that (nice feature! May hijack the concept/name for sluRm). As I said it before, the comparison table was built "to the best of my knowledge". More than happy to keep working on that. Sidenote, perhaps this can be a project of its own, like R4HPC or something. Not sure if there's something like that around, but perhaps having a git(hub) resource for general HPC with R may be useful (thoughts?).
