### 2019-06-24 USCbiostats/sluRm#15 @gvegayon
I really liked this post https://rpubs.com/Jouni_Helske/225931 (mentioned in the issue you pointed out). Following one of the (rather simple) tests, which as you suggested the simple `1:njobs` does not pass, I wrote this simple (sort of arbitrary) function:

```r
# A naive seeding function
new_seeds <- function(nseeds) {

  sum_of_seed <- sum(.Random.seed)
  sum_of_seed <- sum_of_seed - (sum_of_seed %/% 100L)*100L
  seeds <- (1L:nseeds) * 10L
  (seeds + sum_of_seed ^ (1L:nseeds %% 2L))[order(stats::runif(nseeds))]

}

# "Test" from https://rpubs.com/Jouni_Helske/225931
last_rand_seq <- list()
random <- function(kind) {
  set.seed(123)
  RNGkind(kind = kind)
  x <- numeric(1e5)
  seeds <- new_seeds(1e5)
  for(i in seq_along(x)){
    set.seed(seeds[i])
    x[i] <- runif(1)
  }

  last_rand_seq <<- c(last_rand_seq, list(x))

  acf(x, main = kind)
}
kinds <- c("Mersenne-Twister", "Marsaglia-Multicarry", "L'Ecuyer-CMRG")

invisible(sapply(kinds, random)) # Looks nice, no autocorrelation
```

This function is reproducible, i.e., seeds can be reproduced. Will including this function to generate the random seeds instead of using `1:njobs` be enough for JOSS? From what I read in the discussion this is still a problem without a clear solution.
