### 2019-06-24 USCbiostats/sluRm#15 @mllg
I'm not sure I understand how the function `new_seeds()` helps to solve the problem

Your seeds are determined by the current RNG state on which you apply some transformations. Wouldn't you get the same quality of randomness with `as.integer(runif(nseeds, 1, .Machine$integer.max))`?

Either way, I think that the proposed solution (or just calling `runif()`) would be sufficient for me -- the arguably better solution is the one proposed by @HenrikBengtsson in the linked issue.

Note that there are two special cases which you might also need to address to get the seeding right:

1) `.Random.seed` is `NULL` until the RNG is initiated (usually by drawing the first random number).
2) `sum(.Random.seed)` might overflow.
