### 2019-06-24 USCbiostats/sluRm#16 @mllg
> Sidenote, perhaps this can be a project of its own, like R4HPC or something. Not sure if there's something like that around, but perhaps having a git(hub) resource for general HPC with R may be useful (thoughts?).

I'm not aware of something like this. Might be a worthwhile endeavor.
