### 2019-06-25 USCbiostats/sluRm#18 @gvegayon
Traceback is somewhat possible (on the horizon). Right now you can set up debug mode which tells `sbatch` to submit jobs using `sh` instead of `sbatch`, hence you can run the thing on your machine and inspect the log-files afterward. Perhaps, as you suggest, I could try to run the thing interactively instead with a `tryCatch`.

This sounds like a good thing to have (the example itself). I'll work on this for the future.
