### 2019-06-25 91909a7
Addressing #5, #6, #7, #8, #15, and #17
