+++ paper/paper.md
@@ -30,17 +30,25 @@ While the R programming language [@R] has not been developed for HPC settings, t
 
 The `sluRm` R package provides tools for using R in HPC settings that work with Slurm. It provides wrappers and auxiliary functions that allow the user to seamlessly integrate their analysis pipeline with HPC, putting emphasis on providing the user with a family of functions similar to those that the `parallel` R package [@R] provides.
 
-While there are other tools for integrating R in a HPC envirnment that works with Slurm, `sluRm` has some advantages regarding syntax, number of dependencies, and flexibility. 
+While there are other tools for integrating R in a HPC envirnment that works with Slurm--see for example `rslurm`[@Marchand2018], `batchtools` [@Bischl2015], `drake` [@Landau2019], `future.batchtools` [@Bengtsson2019], `clustermq` [@Schubert2019]--`sluRm` has some advantages regarding syntax, number of dependencies, and flexibility (in terms of the integration with Slurm itself). In particular, you may want to use `sluRm` if you:
 
-1.  Compared to `rslurm` [@Marchand2018], this R package's workflow is closer to the `parallel` package. `sluRm` has wrappers such as `Slurm_lapply` and `Slurm_Map` that have a similar syntax to `parLapply` and `parMap` from the `parallel` package; which is not true for `rslurm` which depends on a single function `slurm_apply`.
+1. Need a dependency-free tool. Besides of Slurm itself[^actually], this R package only depends on other R packages that are part of base R.,
 
-2.  Compared to `batchtools` [@Bischl2015], and to `future.batchtools` [@Bengtsson2019] `sluRm` is significantly more lightweight as it has 0 dependencies on R packages other than those shipped with R. `batchtools` and `future.batchtools` have 20 and 24 dependencies each.[^date]
+[^actually]: In fact, users can install this R package regardless of whether they have or they don't have Slurm on their systems. The debug mode of this software allows users to setup jobs (including R scripts and batch files) without having to submit them to a Slurm job-scheduler.
 
-3.  In terms of flexibility, `sluRm`'s API is significantly more flexible compared to the other R packages mentioned. While both `rslurm` and `batchtools` allow the user to provide template files to create personalized job configurations for Slurm, in `sluRm` job configurations are created programatically.
-    
-[^date]: The number of dependencies was obtained from https://tinyverse.netlifly.com, which reports the number of recursive dependencies as of May 1st, 2019.
+2. need an R package that is fully integrated with Slurm, e.g., submitting jobs with an arbitrary set of Slurm parameters without the need of using templates, call Slurm commands from within R like `sacct`, `scancel`, `squeue`, `sbatch`, etc. with their corresponding flags, and
 
-In summary, `sluRm` provides a purpose-built alternative for R users working in a HPC environment with Slurm.
+3. what to use an R package that is ready-to-go. Once loaded, users can submit jobs by just specifying how many cores, for example, they need.
+
+Other features that are included with this R package, and that are available in some others, are:
+
+4. Use a syntax similar to the apply family of functions in the parallel R package, including `Slurm_lapply`, `Slurm_sapply`, `Slurm_EvalQ`, and `Slurm_Map`, 
+
+5. resubmit failed jobs: A very common issue with heterogenous computing clusters is the fact that some jobs succeed while others fail. Partial-job-resubmission is out-of-the-box as users can specify which jobs (as in Job Arrays) should be re-run.
+
+Both of the later two also available in `batchtools`. A comparison table of R packages that work with Slurm is available at https://github.com/USCbiostats/sluRm.
+
+In summary, `sluRm` provides a dependency-free and purpose-built alternative for R users working in a HPC environment with Slurm.
 
 # Funding and Support
 
