+++ paper/paper.bib
@@ -42,12 +42,32 @@ year = {2003}
 }
 
 @Article{Bischl2015,
-  title = {{BatchJobs} and {BatchExperiments}: Abstraction Mechanisms for Using {R} in Batch Environments},
-  author = {Bernd Bischl and Michel Lang and Olaf Mersmann and J{\"o}rg Rahnenf{\"u}hrer and Claus Weihs},
-  journal = {Journal of Statistical Software},
-  year = {2015},
-  volume = {64},
-  number = {11},
-  pages = {1--25},
-  url = {http://www.jstatsoft.org/v64/i11/},
+  title = {batchtools: Tools for R to work on batch systems},
+  author = {Michel Lang and Bernd Bischl and Dirk Surmann},
+  journal = {The Journal of Open Source Software},
+  year = {2017},
+  month = {feb},
+  volume = {2},
+  number = {10},
+  doi = {10.21105/joss.00135},
+  url = {https://doi.org/10.21105/joss.00135},
 }
+
+@Article{Schubert2019,
+  title = {clustermq enables efficient parallelisation of genomic analyses},
+  author = {Michael Schubert},
+  journal = {Bioinformatics},
+  month = {May},
+  year = {2019},
+  language = {en},
+  doi = {10.1093/bioinformatics/btz284},
+  url = {https://github.com/mschubert/clustermq},
+}
+
+@Manual{Landau2019,
+    title = {drake: A Pipeline Toolkit for Reproducible Computation at Scale},
+    author = {William Michael Landau},
+    year = {2019},
+    note = {R package version 7.4.0},
+    url = {https://CRAN.R-project.org/package=drake},
+  }
