+++ R/Map.R
@@ -37,13 +37,13 @@ Slurm_Map <- function(
   f,
   ...,
   njobs       = 2L,
-  mc.cores    = getOption("mc.cores", 2L),
+  mc.cores    = 1L,
   job_name    = opts_sluRm$get_job_name(),
   tmp_path    = opts_sluRm$get_tmp_path(),
   plan        = "collect",
   sbatch_opt  = list(ntasks=1L, `cpus-per-task`=mc.cores),
   rscript_opt = list(vanilla=TRUE),
-  seeds       = 1L:njobs,
+  seeds       = NULL,
   compress    = TRUE,
   export      = NULL,
   libPaths    = .libPaths(),
