+++ R/options.R
@@ -289,6 +289,7 @@ print.opts_sluRm <- function(x, ...) {
 }
 
 .onLoad <- function(libname, pkgname) {
+
   opts_sluRm$set_tmp_path(getwd())
 
   tmp <- tempfile("sluRm-job-", opts_sluRm$get_tmp_path())
