### 2019-06-25 4efeed9
Rebuilding website and re-running checks after merging the chdir branch (addresses #14)
