### 2019-06-25 USCbiostats/sluRm#14 @gvegayon
This actually turned to be a big issue (required me to make a new branch). I was kind of using `chdir` as a tmp_dir, now changed and it shouldn't be an issue anymore.
