### 2019-06-25 USCbiostats/sluRm#13 @gvegayon
Addressed in 91909a7c4fd4b94811d473384b56712285e7baa5
