### 2019-06-25 USCbiostats/sluRm#9 @gvegayon
No, it is not the intended behavior. You should get a warning when overwriting a job. I'll address this after the review.
