### 2019-06-25 3996c42
Solving small couple of bugs introduced after changes in chdir
