### 2019-06-25 #1493 @gvegayon
OK @mschubert @mllg, I've worked on the issues you created on the project's repo. Addressed all of the required ones plus some of those which were not mandatory (thanks!). LMK if you are OK with the changes, and if so, please close the issues that you consider complete.

Also, following recent changes on how to list contributions on R packages (if you are OK with it) I've added both of you in the list of authors with the role of `"rev"` (see [here](https://github.com/USCbiostats/sluRm/blob/91909a7c4fd4b94811d473384b56712285e7baa5/DESCRIPTION#L8-L9)).

PS: A thing that I'm not sure whether JOSS does but I think is worthy is to keep track of how much projects change after submission. I am proud to say that my `git diff --stat` is (+)917 (-)269 lines since the package was submitted (on R/Rmd/md files), so yes, this review has required me to make significant changes towards I think is an improved version of the package, in other words, I think the review process works!
