### 2019-06-25 fcd38cb
Updating vignette and fixing bug in sacct when the ouput was of only 1 row [skip ci].
