### 2019-06-25 #1493 @karthik
Thanks @gvegayon! I'm glad you added the reviewers with the appropriate MARC roles (it's recommended at rOpenSci but not here). If either reviewer objects or declines, you can revert the change.

Glad to see the process working! Please see another issue I raised on your repo.
