### 2019-06-26 3eca011
Addressing #19 [skip ci]
