+++ man-roxygen/slurm.R
@@ -1,9 +1,9 @@
 #' @param job_name Character. Name of the job to be passed to `Slurm`.
 #' @param tmp_path Character. Path to the directory where all the data (including
-#' scripts) will be stored. Notice that this path must be accesible by all the
+#' scripts) will be stored. Notice that this path must be accessible by all the
 #' nodes in the network (See [opts_sluRm]).
 #' @param plan A character scalar. (See [the_plan]).
-#' @param njobs Integer. Number of jobs to specity.
+#' @param njobs Integer. Number of jobs to use in the job-array.
 #' @param sbatch_opt,rscript_opt List. Options to be passed via flags to
 #' the bash file as `#SBATCH` and to `Rscript` respectively.
 #' @param compress Logical scalar (default `TRUE`). Passed to [saveRDS]. Setting
