### 2019-06-26 b0a485b
Adding news and other features for docs/* [skip ci]
