### 2019-06-26 #1493 @karthik
Everything LGTM.

Over you to @openjournals/joss-eics
