### 2019-06-26 #1493 @mschubert
`drake` citation looks wrong, shouldn't this be JOSS instead of the package URL?
