### 2019-06-26 #1493 @kyleniemeyer
Yes, @gvegayon, please update the drake citation to the JOSS paper:
```
@article{MichaelLandau2018,
  doi = {10.21105/joss.00550},
  year = {2018},
  month = jan,
  volume = {3},
  number = {21},
  pages = {550},
  author = {William Michael Landau},
  title = {The drake {R} package: a pipeline toolkit for reproducibility and high-performance computing},
  journal = {The Journal of Open Source Software}
}```
