### 2019-06-26 3c12943
Correct drake citation [skip ci]
