+++ paper/paper.bib
@@ -64,10 +64,14 @@ year = {2003}
   url = {https://github.com/mschubert/clustermq},
 }
 
-@Manual{Landau2019,
-    title = {drake: A Pipeline Toolkit for Reproducible Computation at Scale},
-    author = {William Michael Landau},
-    year = {2019},
-    note = {R package version 7.4.0},
-    url = {https://CRAN.R-project.org/package=drake},
-  }
+@article{Landau2018,
+  doi = {10.21105/joss.00550},
+  year = {2018},
+  month = jan,
+  volume = {3},
+  number = {21},
+  pages = {550},
+  author = {William Michael Landau},
+  title = {The drake {R} package: a pipeline toolkit for reproducibility and high-performance computing},
+  journal = {The Journal of Open Source Software}
+}
