### 2019-06-26 #1493 @gvegayon
> `drake` citation looks wrong, shouldn't this be JOSS instead of the package URL?

Yes @mschubert, thanks @kyleniemeyer, I looked for papers related to drake on google scholar without success (perhaps needed to look harder?). Changed.

@karthik I see on the bottom of the pdf version is skipping part of my last name (is *Vega Yon*, not just *Yon*). I know this may be a bit painful, but I was wondering if there's any way to fix that? If not, perhaps a PR to JOSS to follow an approach as the `person(given=, family=,middle=)` function in R (for the future, of course).
