### 2019-06-26 #1493 @kyleniemeyer
@gvegayon oh, good catch on your name.

@arfon, is there a way to ensure whedon captures compound / double surnames? Would wrapping it in `{ }` work, as in BibTeX?
