### 2019-06-27 #1493 @arfon
> Thanks @kyleniemeyer and @gvegayon for catching this. Let's wait for @arfon to fix this manually.

Yes, this will have to be fixed manually at the end of the review. From a quick inspection it looks like we're not done with this review yet? If so, I'll fix this once the paper is ready to be accepted.
