### 2019-06-27 #1493 @kyleniemeyer
Ah, I didn't notice the checklists.

@mschubert and @mllg, can you confirm that all the outstanding checklist items are satisfied? (And check them off if so)
