### 2019-06-28 #1493 @mschubert
Looks good to me!

I haven't seen any comments from @pmarjora, is it customary for JOSS that he also confirms his authorship?

Last minor comments: Typos in the paper (resubmission, "the later" should be "the latter") and README ("Whereas is reporting" doesn't make sense, did you mean "Whether it is"?)
