### 2019-06-28 ec746cb
Minor issues on paper and readme (openjournals/joss-reviews#1493) + adding examples.
