+++ man/Slurm_collect.Rd
@@ -20,3 +20,14 @@ of whether the job is completed or not.}
 \description{
 Collect the results
 }
+\examples{
+\dontrun{
+# Collecting a job after calling it
+job <- Slurm_EvalQ(sluRm::WhoAmI(), njobs = 4, plan = "wait")
+Slurm_collect(job)
+
+# Collecting a job from a previous R session
+job <- read_slurm_job("/path/to/a/job/tmp_dir")
+Slurm_collect(job)
+}
+}
