### 2019-06-28 ac76c7e
Fixing small bug when parsing flags. @pmarjora is (ctb,ths) now.
