+++ DESCRIPTION
@@ -3,7 +3,7 @@ Title: A Lightweight Wrapper for Slurm
 Version: 0.1-0
 Authors@R: c(
   person("George", "Vega Yon", email = "g.vegayon@gmail.com", role = c("aut", "cre"), comment = c(ORCID = "0000-0002-3171-0844")),
-  person("Paul", "Marjoram", email = "pmarjora@usc.edu", role = c("aut"), comment = c(ORCID = "0000-0003-0824-7449")),
+  person("Paul", "Marjoram", email = "pmarjora@usc.edu", role = c("ctb", "ths"), comment = c(ORCID = "0000-0003-0824-7449")),
   person("National Cancer Institute (NCI)", role = "fnd", comment = "Grant Number 5P01CA196569-02"),
   person("Michael", "Schubert", role = "rev", comment = c(what = "JOSS reviewer", ORCID = "0000-0002-6862-5221")),
   person("Michel", "Lang", role = "rev", comment = c(what = "JOSS reviewer", ORCID = "0000-0001-9754-0393"))
