### 2019-06-30 #1493 @karthik
> I haven't seen any comments from @pmarjora, is it customary for JOSS that he also confirms his authorship?

We don't require every author to confirm separately.
