### 2019-06-30 #1493 @mllg
The issues I raised have been adequately addressed. From my side, this is ready for publication.
