### 2019-07-03 #1493 @karthik
Looks like both reviewers have now properly signed off. Over to you again @openjournals/joss-eics
