### 2019-07-04 #1493 @labarba
👋  @gvegayon — The JOSS publication process requires that after all revisions to the software and paper are complete and the reviewers and editor recommend publication, you do the following:

- make a tagged release of your software, and report the version number here so we can update the metadata
- make a full deposit of your software in an archive that provides persistence and a global identifier such as a DOI (Zenodo is the most popular, but other archives with similar services exist and are permitted); report the DOI of the software archive here, so we can update the metadata.

You may find this helpful: https://guides.github.com/activities/citable-code/
