### 2019-07-08 4497fb0
Bump version for JOSS review (see https://github.com/openjournals/joss-reviews/issues/1493#issuecomment-508407191) [skip ci]
