+++ NEWS.md
@@ -1,3 +1,7 @@
+# sluRm 0.2-0
+
+* JOSS review (finalizing).
+
 # sluRm 0.1-0
 
 * Added a `NEWS.md` file to track changes to the package.
