### 2019-07-08 #1493 @labarba
I found the tag of the version https://github.com/USCbiostats/sluRm/releases
