### 2019-07-08 #1493 @gvegayon
> And the DOI unfortunately does not resolve http://doi.org/10.5281/zenodo.3272773

Now it does :)
