### 2019-07-08 #1493 @gvegayon
> Please check the Zenodo deposit has the correct metadata, this includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it); you may also add the authors' ORCID.

Just did: Authors (with ORCID), title and description (as provided in the R package). Question: On the **upload type** field, should I check **Publication** or **Software**?
