### 2019-07-08 #1493 @labarba
Big thanks to our editor: @karthik, and reviewers: @mschubert, @mllg —JOSS runs on your contribution 🙏
