### 2019-07-08 #1493 @gvegayon
Great! Thanks :)! one small problem! my last name is "Vega Yon", not "Yon" (see [here](https://github.com/openjournals/joss-reviews/issues/1493#issuecomment-505993965))
