### 2019-07-08 #1493 @arfon
> Ping @arfon for this ☝️

This should be fixed now. The PDF on the JOSS site can take a while to refresh as it's cached.
