### 2019-10-08 #1493 @arfon
> Hi @arfon I've finished rebranding sluRm to slurmR (paper included). The version of the repository that should be published is here: https://github.com/USCbiostats/slurmR/tree/257577f2197747ea926dfc9a7349c795d7a85e88

Roger that. The paper should now be updated although it might take a few hours to show up modified on the JOSS site as the PDFs are cached.
