### 2021-08-12 #1493 @gvegayon
Hey @arfon! I realized that my [last name has not been fixed for the citation](https://github.com/openjournals/joss-reviews/issues/1493#issuecomment-509410314). No rush, but I would appreciate if you could take a look :D!
